Refactor out manager tests to separate files for ease of pytest setup
