+++ letta/schemas/user.py
@@ -3,6 +3,7 @@ from typing import Optional
 
 from pydantic import Field
 
+from letta.constants import DEFAULT_ORG_ID
 from letta.schemas.letta_base import LettaBase
 
 
@@ -20,10 +21,8 @@ class User(UserBase):
         created_at (datetime): The creation date of the user.
     """
 
-    id: str = UserBase.generate_id_field()
-    organization_id: Optional[str] = Field(
-        ..., description="The organization id of the user"
-    )  # TODO: dont make optional, and pass in default org ID
+    id: str = Field(..., description="The id of the user.")
+    organization_id: Optional[str] = Field(DEFAULT_ORG_ID, description="The organization id of the user")
     name: str = Field(..., description="The name of the user.")
     created_at: datetime = Field(default_factory=datetime.utcnow, description="The creation date of the user.")
     updated_at: datetime = Field(default_factory=datetime.utcnow, description="The update date of the user.")
